requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-Session==0.5.0
requests==2.32.5

